JIT becomes REPL
The orc JIT support turns into a REPL by emitting llvm modules
incrementally which allows them to be passed to ORC concurrently with
the code being typed.

Normally, we output a giant llvm module containing the code for the
whole application but this does not work for ORC since modules are
compiled incrementally and not the full application is not known ahead
of time.

In REPL mode, we'll finalize the llvm module every time there is
top-level code to be executed (that would have been called from `main`
in an ordinary compile) and pass it to ORC, continuing code generation
in a new llvm module.

The old JIT mode becomes a special case where we do the above at the end
of compilation instead of incrementally.

* move `init` to LLGen - this might slightly change init order of some
generic globals but didn't seem to break anything
* compile JIT code with linkonce linkage and reference it from follow-up
repl code
* fix dead block generation to not use global context
* generate forwarded proc's for every closed module
* generate global destructors in main module only